refactor : 기능별로 api 요청 코드 분리
