+++ client/api/makePostRequest.ts
@@ -1,0 +1,5 @@
+import axios from 'axios';
+
+export const makePostRequest = async (endpoint: string, data: object) => {
+  return await axios.post(endpoint, data);
+};
